+++ package.json
@@ -61,6 +61,7 @@
     "@rtm/shared-lib": "*",
     "@rtm/shared-types": "*",
     "@vercel/speed-insights": "^1.0.2",
+    "base64url": "^3.0.1",
     "bentocache": "1.0.0-beta.7",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.0.0",
