+++ src/components/pages/blog/helpers/constants.ts
@@ -1,5 +1,1 @@
 export const FILTERS_KEY = 'tags';
-export const PAGE_KEY = 'page';
-
-export const FIRST_PAGE_IDX = 1;
-export const MIN_PAGES_AMOUNT = 1;
