+++ src/components/ui/helpers/PaginatedElements/constants.ts
@@ -1,0 +1,3 @@
+export const PAGE_KEY = 'page';
+export const FIRST_PAGE_IDX = 1;
+export const MIN_PAGES_AMOUNT = 1;
